feat: implement Spring Data JPA query method lexer with context-aware parsing
- Add QueryMethodLexer with entity metadata validation
  - Context-aware tokenize(Class<?>, String) validates property paths
  - Supports nested properties (department.address.city)
  - Supports @OneToOne, @ManyToOne, @Embedded relationships
  - Supports self-referential relationships
  - Fixed number handling to prevent false matches (age2 vs age)

- Add QueryPlanner with state machine token processing
  - Processes tokens: property → operator → modifier → condition
  - Supports 20+ operators including new ones (NOT_LIKE, IS_NULL, etc.)
  - Handles IgnoreCase as separate operator token
  - Supports AND/OR combinators and OrderBy

- Add LogicalQuery model with 24 operators
  - ReturnKind: ONE_OPTIONAL, MANY_LIST, EXISTS_BOOL, COUNT_LONG
  - Operators: EQ, NE, GT, GTE, LT, LTE, IN, NOT_IN, BETWEEN,
    LIKE, NOT_LIKE, CONTAINING, STARTING_WITH, ENDING_WITH,
    IS_NULL, NOT_NULL, IS_TRUE, IS_FALSE, BEFORE, AFTER,
    IGNORE_CASE_EQ, IGNORE_CASE_LIKE

- Add RepositoryRuntime for query execution
  - Typed entrypoints: list0/1/2, optional1, exists1, count0/1/2
  - O(1) queryId lookup with pre-compiled conditions
  - Zero reflection in hot path

- Add RepositoryEmitter using ByteBuddy
  - Generates repository implementation at runtime
  - Uses MethodDelegation with @RuntimeType for flexibility
  - INJECTION classloading strategy to avoid isolation

- Add 53 tests covering all features
  - ContextAwareLexerTest: 15 tests
  - ComplexNestingTest: 23 tests (deep nesting, self-referential, embedded)
  - QueryPlannerIntegrationTest: 15 tests

- Fix JaCoCo configuration
  - Use propertyName=argLine to match surefire configuration
  - Remove includes/excludes filters that were blocking coverage

- Remove old QueryMethodParser and RepositoryBytecodeGenerator
  - Replaced by new QueryMethodLexer + QueryPlanner architecture
  - 53 new tests replace 66 old incompatible tests

+++ memris-core/src/main/java/io/memris/spring/TypeCodes.java
@@ -9,13 +9,13 @@ public final class TypeCodes {
         // Utility class - no instances
     }
 
-    public static final int TYPE_INT = 0;
-    public static final int TYPE_LONG = 1;
-    public static final int TYPE_BOOLEAN = 2;
-    public static final int TYPE_BYTE = 3;
-    public static final int TYPE_SHORT = 4;
-    public static final int TYPE_FLOAT = 5;
-    public static final int TYPE_DOUBLE = 6;
-    public static final int TYPE_CHAR = 7;
-    public static final int TYPE_STRING = 8;
+    public static final byte TYPE_INT = 0;
+    public static final byte TYPE_LONG = 1;
+    public static final byte TYPE_BOOLEAN = 2;
+    public static final byte TYPE_BYTE = 3;
+    public static final byte TYPE_SHORT = 4;
+    public static final byte TYPE_FLOAT = 5;
+    public static final byte TYPE_DOUBLE = 6;
+    public static final byte TYPE_CHAR = 7;
+    public static final byte TYPE_STRING = 8;
 }
